+++ .dir-locals.el
@@ -1,0 +1,5 @@
+((haskell-mode . ((haskell-indent-spaces . 4)
+                  (hindent-style . "johan-tibell")
+                  (haskell-process-path-ghci . "stack")
+                  (haskell-process-args-ghci . ("ghci" "--with-ghc=ghci-ng"))
+                  (haskell-process-type . ghci))))